Consolidate test sweepers. Add faux test to force all sweepers to run.
