+++ okta/app_test.go
@@ -1,9 +1,6 @@
 package okta
 
 import (
-	"context"
-	"fmt"
-	"strings"
 	"testing"
 )
 
@@ -33,27 +30,3 @@ func TestLogoStateFunc(t *testing.T) {
 	}
 }
 
-func deleteTestApps(client *testClient) error {
-	appList, err := listApps(context.Background(), client.oktaClient, &appFilters{LabelPrefix: testResourcePrefix}, defaultPaginationLimit)
-	if err != nil {
-		return err
-	}
-	var warnings []string
-	for _, app := range appList {
-		warn := fmt.Sprintf("failed to sweep an application, there may be dangling resources. ID %s, label %s", app.Id, app.Label)
-		_, err := client.oktaClient.Application.DeactivateApplication(context.Background(), app.Id)
-		if err != nil {
-			warnings = append(warnings, warn)
-		}
-		resp, err := client.oktaClient.Application.DeleteApplication(context.Background(), app.Id)
-		if is404(resp) {
-			warnings = append(warnings, warn)
-		} else if err != nil {
-			return err
-		}
-	}
-	if len(warnings) > 0 {
-		return fmt.Errorf("sweep failures: %s", strings.Join(warnings, ", "))
-	}
-	return nil
-}
