+++ okta/idp_test.go
@@ -2,8 +2,6 @@ package okta
 
 import (
 	"context"
-
-	"github.com/okta/okta-sdk-golang/v2/okta/query"
 )
 
 func createDoesIdpExist() func(string) (bool, error) {
@@ -13,23 +11,3 @@ func createDoesIdpExist() func(string) (bool, error) {
 	}
 }
 
-func deleteTestIdps(client *testClient) error {
-	providers, _, err := client.oktaClient.IdentityProvider.ListIdentityProviders(context.Background(), &query.Params{Q: "testAcc_"})
-	if err != nil {
-		return err
-	}
-	for _, idp := range providers {
-		_, err := client.oktaClient.IdentityProvider.DeleteIdentityProvider(context.Background(), idp.Id)
-		if err != nil {
-			return err
-		}
-
-		if idp.Type == saml2Idp {
-			_, err := client.oktaClient.IdentityProvider.DeleteIdentityProviderKey(context.Background(), idp.Protocol.Credentials.Trust.Kid)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
