+++ okta/resource_okta_group_test.go
@@ -8,24 +8,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/okta/okta-sdk-golang/v2/okta/query"
 )
-
-func sweepGroups(client *testClient) error {
-	var errorList []error
-	// Should never need to deal with pagination, limit is 10,000 by default
-	groups, _, err := client.oktaClient.Group.ListGroups(context.Background(), &query.Params{Q: testResourcePrefix})
-	if err != nil {
-		return err
-	}
-
-	for _, s := range groups {
-		if _, err := client.oktaClient.Group.DeleteGroup(context.Background(), s.Id); err != nil {
-			errorList = append(errorList, err)
-		}
-	}
-	return condenseError(errorList)
-}
 
 func TestAccOktaGroup_crud(t *testing.T) {
 	ri := acctest.RandInt()
