+++ okta/resource_okta_inline_hooks_test.go
@@ -3,36 +3,11 @@ package okta
 import (
 	"context"
 	"net/http"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
-
-func sweepInlineHooks(client *testClient) error {
-	var errorList []error
-	hooks, _, err := client.oktaClient.InlineHook.ListInlineHooks(context.Background(), nil)
-	if err != nil {
-		return err
-	}
-	for _, hook := range hooks {
-		if !strings.HasPrefix(hook.Name, testResourcePrefix) {
-			continue
-		}
-		if hook.Status == statusActive {
-			_, _, err = client.oktaClient.InlineHook.DeactivateInlineHook(context.Background(), hook.Id)
-			if err != nil {
-				errorList = append(errorList, err)
-			}
-		}
-		_, err = client.oktaClient.InlineHook.DeleteInlineHook(context.Background(), hook.Id)
-		if err != nil {
-			errorList = append(errorList, err)
-		}
-	}
-	return condenseError(errorList)
-}
 
 func TestAccOktaInlineHook_crud(t *testing.T) {
 	ri := acctest.RandInt()
