+++ okta/resource_okta_link_definition_test.go
@@ -3,28 +3,11 @@ package okta
 import (
 	"context"
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
-
-func sweepLinkDefinitions(client *testClient) error {
-	var errorList []error
-	linkedObjects, _, err := getOktaClientFromMetadata(testAccProvider.Meta()).LinkedObject.ListLinkedObjectDefinitions(context.Background())
-	if err != nil {
-		return err
-	}
-	for _, object := range linkedObjects {
-		if strings.HasPrefix(object.Primary.Name, testResourcePrefix) {
-			if _, err := getOktaClientFromMetadata(testAccProvider.Meta()).LinkedObject.DeleteLinkedObjectDefinition(context.Background(), object.Primary.Name); err != nil {
-				errorList = append(errorList, err)
-			}
-		}
-	}
-	return condenseError(errorList)
-}
 
 func TestAccOktaLinkDefinition(t *testing.T) {
 	ri := acctest.RandInt()
