+++ okta/resource_okta_policy_rule_sign_on_test.go
@@ -7,12 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/okta/terraform-provider-okta/sdk"
 )
-
-func deleteSignOnPolicyRules(client *testClient) error {
-	return deletePolicyRulesByType(sdk.SignOnPolicyType, client)
-}
 
 func TestAccOktaPolicyRuleSignon_defaultErrors(t *testing.T) {
 	config := testOktaPolicyRuleSignOnDefaultErrors(acctest.RandInt())
