+++ okta/resource_okta_user_type_test.go
@@ -4,26 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
-
-func sweepUserTypes(client *testClient) error {
-	userTypeList, _, _ := client.oktaClient.UserType.ListUserTypes(context.Background())
-	var errorList []error
-	for _, ut := range userTypeList {
-		if strings.HasPrefix(ut.Name, testResourcePrefix) {
-			if _, err := client.oktaClient.UserType.DeleteUserType(context.Background(), ut.Id); err != nil {
-				errorList = append(errorList, err)
-			}
-		}
-	}
-	return condenseError(errorList)
-}
 
 func TestAccOktaUserType_crud(t *testing.T) {
 	ri := acctest.RandInt()
